Create ERF entry for NSHM23 WUS
